Modify update.sh to never remove plugins
Deleting bundles which are git repos itself is unnecessary.
Simply update git repos instead.

+++ update.sh
@@ -3,6 +3,8 @@
 # Usage: ./update.sh [pattern]
 #
 # Specify [pattern] to update only repos that match the pattern.
+#
+# When deleting plugins, uncomment here and delete in bundles directory
 
 repos=(
 
@@ -42,13 +44,12 @@ repos=(
 set -e
 dir=~/.dotfiles/.vim/bundle
 
-if [ -d $dir -a -z "$1" ]; then
-  temp="$(mktemp -d -t bundleXXXXX)"
-  echo "▲ Moving old bundle dir to $temp"
-  mkdir -p $temp
-  mv $dir/* "$temp/" || true
+if [ ! -d $dir ]
+then
+  mkdir $dir
 fi
 
+wd=`pwd`
 for repo in ${repos[@]}; do
   if [ -n "$1" ]; then
     if ! (echo "$repo" | grep -i "$1" &>/dev/null) ; then
@@ -58,11 +59,18 @@ for repo in ${repos[@]}; do
   plugin="$(basename $repo | sed -e 's/\.git$//')"
   [ "$plugin" = "vim-styled-jsx" ] && plugin="000-vim-styled-jsx" # https://goo.gl/tJVPja
   dest="$dir/$plugin"
-  rm -rf $dest
   (
-    git clone --depth=1 -q https://github.com/$repo $dest
-    rm -rf $dest/.git
-    echo "· Cloned $repo"
-  ) &
+    if [ -d $dest ]
+    then
+      cd $dest
+      git pull
+      cd $wd
+      echo ". Updated $repo"
+    else
+      rm -rf $dest
+      git clone --depth=1 -q https://github.com/$repo $dest
+      echo "· Cloned $repo"
+    fi
+   ) &
 done
 wait
